requirements.txt
torch>=2.0.1
transformers==4.31.0
datasets==2.13.0
tqdm>=4.65.0
pyyaml>=6.0
accelerate>=0.21.0
wandb>=0.15.5
